Casting, id type, naming & missing include fixes
1. Change unsigned long long to uint64_t
2. Use numeric_limits<uint64_t>::max() to get the max value
3. Use static_cast instead of c-style cast to invoke conversion function
4. Change persist_id to persistId as per naming convention
5. Use uint64_t for id in test cases, otherwise it will fail
6. Remove extra inclusion of header <random>

+++ alica_tests/src/test/test_agent_discovery.cpp
@@ -60,7 +60,7 @@ TEST_F(AlicaEngineAgentDiscoveryTest, AgentDiscovered)
     ASSERT_EQ(2, aes[0]->getTeamManager().getActiveAgentIds().size());
     ASSERT_EQ(2, aes[1]->getTeamManager().getActiveAgentIds().size());
 
-    int id = 8;
+    uint64_t id = 8;
     const alica::Agent* hairydiscovered = aes[0]->getTeamManager().getAgentByID(aes[0]->getId(id));
     const alica::Agent* hairyoriginal = aes[1]->getTeamManager().getLocalAgent();
     verifyAgents(hairydiscovered, hairyoriginal);
